requirements.txt
flask
boto3
torch
torchvision
pillow
python-dotenv
gunicorn